+++ extension-style/src/main/java/com/mapbox/maps/extension/style/sources/SourceExt.kt
@@ -17,11 +17,11 @@ import com.mapbox.maps.extension.style.utils.silentUnwrap
 fun StyleManagerInterface.getSource(sourceId: String): Source? {
   return this.getStyleSourceProperty(sourceId, "type").silentUnwrap<String>()?.let { type ->
     when (type) {
-      "vector" -> VectorSource.Builder(sourceId).build().also { it.delegate = this }
-      "geojson" -> GeoJsonSource.Builder(sourceId).build().also { it.delegate = this }
-      "image" -> ImageSource.Builder(sourceId).build().also { it.delegate = this }
-      "raster-dem" -> RasterDemSource.Builder(sourceId).build().also { it.delegate = this }
-      "raster" -> RasterSource.Builder(sourceId).build().also { it.delegate = this }
+      "vector" -> VectorSource.Builder(sourceId).build().also { it.delegate = this as StyleInterface }
+      "geojson" -> GeoJsonSource.Builder(sourceId).build().also { it.delegate = this as StyleInterface }
+      "image" -> ImageSource.Builder(sourceId).build().also { it.delegate = this as StyleInterface }
+      "raster-dem" -> RasterDemSource.Builder(sourceId).build().also { it.delegate = this as StyleInterface }
+      "raster" -> RasterSource.Builder(sourceId).build().also { it.delegate = this as StyleInterface }
       else -> {
         Logger.e("StyleSourcePlugin", "Source type: $type unknown.")
         null